Should of put this in multiple commits, oh well.
Flushed out the canon class.  It can now fire canon balls properly.

Fixed the CameraEdgeDetector so that ships cannot accidentally slip away.

Added a little ship splash animation so you can tell they are kinda moving.  This is all temp art so it will all be thrown out but it makes a nice effect for the prototype.

Created the Ship class that will hold a reference to all the individual ship parts and will be responsible for telling all the parts when events happen (collision with canon ball, collision with other ship, etc)