Added a editor folder.  Will be used for all editor inspectors and should therefor follow the same folder structure as scripts.
Added a working folder and a text file inside it.  This folder is a build location and the file inside describes a naming convention for builds to standardize it.  By default everything inside the working folder is git ignored.

Changed the default size of the game so that it is 16:9.